server/requirements.txt
fastapi
uvicorn
pandas
python-multipart
panderaannotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.0
h11==0.16.0
idna==3.10
mypy-extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pandera==0.25.0
psycopg2==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typeguard==4.4.4
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.35.0
